feat: clang-tidy fixes

+++ lib/libc/time/difftime.c
@@ -1,6 +1,6 @@
-#include <time.h>
+#include <time.h> // for time_t, difftime
 
 double difftime(time_t time1, time_t time0)
 {
-	return time1 - time0;
+	return (double)(time1 - time0);
 }
