+++ lib/libc/unistd/isatty.c
@@ -1,4 +1,4 @@
-#include "asm/unistd_64.h" // for __NR_ioctl
+
 
 #include <asm-generic/ioctls.h> // for TIOCGWINSZ
 #include <syscall.h>		// for __syscall_3, syscall
@@ -8,5 +8,5 @@
 int isatty(int fildes)
 {
 	struct winsize winsize;
-	return 1 + syscall(ioctl, fildes, TIOCGWINSZ, &winsize);
+	return (int)(1 + syscall(ioctl, fildes, TIOCGWINSZ, &winsize));
 }
